+++ packages/polar/tests/pie/props.ts
@@ -1,0 +1,18 @@
+const pieData = [
+    {
+        id: 'alpha',
+        data: 30,
+    },
+    {
+        id: 'beta',
+        data: 20,
+    },
+    {
+        id: 'gamma',
+        data: 10,
+    },
+]
+
+export const pieProps = {
+    data: pieData,
+}